Update ARM Template deployment method
Updated the ARM Template so that it can be deployed via the "Deploy to Azure" button. Some times compromises have to be made and I've figured that it is better to get the code deployed and working automated, rather than trying to get everything automated. The publish profile for the Function App was proving a bit tricky so this may be revisited in the future.